Translations: Avoid dot in .ts file names
Qt qmake feature 'lrelease' cannot deal with file names containing more than
one dot. The script tries to cut off the part behind the first dot and replace
it with a new extension, assuming the only dot in the file name is the
extension separator. This results in failure to generate installation targets
for translation files. This has been reported as QTBUG-82107.

Although the assumption about number of dots seems to be mistaken, it is still
easier at this point to adapt to the situation rather than fix qmake in all
target distros.